third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDistantLightElement-dom-elevation-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'elevation' attribute of the SVGFEDistantLightElement object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS distantLight.getAttribute('elevation') is "50"
PASS distantLight.getAttribute('elevation') is "20"
PASS successfullyParsed is true

TEST COMPLETE
